refactor: replace SavingsPlanRoutes with SavingsAiRoutes for improved routing structure
